backend/requirements.txt
django==5.2
djangorestframework==3.16.0
nltk==3.9.1
scikit-learn==1.5.0
joblib==1.5.0
numpy==1.26.4
pandas==2.2.0
python-docx==1.1.0
PyPDF2==3.0.1
textract==1.6.5
chardet==5.2.0
tqdm==4.67.1
pylint==3.0.5
django-cors-headers==4.3.0 